django_backend/requirements.txt
Django==5.2.6
psycopg2-binary==2.9.10
celery==5.5.3
redis>=6.0,<7.0
djangorestframework==3.15.2
djangorestframework-simplejwt>=4
django-filter>=25